Created the base for the program
I created two classes.

The first is class is MainWindow. The MainWindow class holds everything
in the program will handle user input.

The second class is called DrawArea. The DrawArea class controls drawing
the lines representing the robot's path. I also plan on having this
class record the robot's path.

As of now, there is only a button and an instance of the DrawArea class
in MainWindow. When the button is pressed, a line is drawn in the draw
area.

Now I am going to go to sleep.